7406 - Turn mc-pgrep into an application
Turn the basic mc-pgrep into an mco application.  As with mc-pgrep the
kill and pkill actions are not represented in the application as they
are just too dangerous for general use.

The timeout in the agent did not match up with the DDL, this resolves
that avoiding short timeout errors on slow machines

The mc-pgrep script was removed